simplify the object wrapper system immensely
i realised that, after +registerWrappedGType: is called, the Class will
already know what its associated GType is, so there's no reason to require
the wrapper to pass it again when init-ing an object instance.  this way,
things should work much better, and the API can be a lot more objc-like,
since -init actually works, and just chains to -initWithProperties:, which
is our new designated initialiser.

now, the only places GType or GObject is exposed in the API are
+registerWrappedGType: and -gobjectPointer.